requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
accelerate==1.9.0
    # via -r requirements.in
blinker==1.9.0
    # via flask
certifi==2025.7.14
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via flask
colorama==0.4.6
    # via
    #   click
    #   tqdm
diffusers==0.34.0
    # via -r requirements.in
filelock==3.18.0
    # via
    #   diffusers
    #   huggingface-hub
    #   torch
    #   transformers
flask==3.1.1
    # via -r requirements.in
fsspec==2025.7.0
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.33.4
    # via
    #   accelerate
    #   diffusers
    #   tokenizers
    #   transformers
idna==3.10
    # via requests
importlib-metadata==8.7.0
    # via diffusers
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   torch
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.6
    # via
    #   accelerate
    #   diffusers
    #   transformers
packaging==25.0
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
pillow==11.3.0
    # via
    #   -r requirements.in
    #   diffusers
psutil==7.0.0
    # via accelerate
pyyaml==6.0.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via
    #   diffusers
    #   transformers
requests==2.32.4
    # via
    #   diffusers
    #   huggingface-hub
    #   transformers
safetensors==0.5.3
    # via
    #   accelerate
    #   diffusers
    #   transformers
sympy==1.14.0
    # via torch
tokenizers==0.21.2
    # via transformers
torch==2.7.1
    # via
    #   -r requirements.in
    #   accelerate
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.53.2
    # via -r requirements.in
typing-extensions==4.14.1
    # via
    #   huggingface-hub
    #   torch
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via flask
zipp==3.23.0
    # via importlib-metadata
